fix more

+++ Classes/Hmdq/HmdqOutput.cs
@@ -561,7 +561,7 @@ namespace SCVRPatcher {
 
         [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
         [JsonPropertyName("Prop_ImuFactoryGyroScale_Vector3")]
-        public virtual List<long> PropImuFactoryGyroScaleVector3 { get; set; }
+        public virtual List<double> PropImuFactoryGyroScaleVector3 { get; set; }
 
         [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
         [JsonPropertyName("Prop_ImuFactoryAccelerometerBias_Vector3")]
